Move AmendementSaisi event to Visam models
