+++ repondeur/zam_repondeur/models/events/amendement.py
@@ -49,22 +49,6 @@ class AmendementEvent(Event):
         )
 
 
-class AmendementSaisi(AmendementEvent):
-    __mapper_args__ = {"polymorphic_identity": "amendement_saisi"}
-
-    icon = "edit"
-
-    summary_template = Template(
-        f"L’amendement a été saisi par <abbr title='$email'>$user</abbr>."
-    )
-
-    def apply(self) -> None:
-        pass
-
-    def render_details(self) -> str:
-        return ""
-
-
 class AmendementRectifie(AmendementEvent):
     __mapper_args__ = {"polymorphic_identity": "amendement_rectifie"}
