+++ repondeur/zam_repondeur/visam/models/__init__.py
@@ -1,1 +1,2 @@
 from .conseil import Conseil, Formation  # noqa
+from .events import AmendementSaisi  # noqa
